Time: 1 ms (25.56%), Space: 47.2 MB (5.41%) - LeetHub

+++ 2529-maximum-count-of-positive-integer-and-negative-integer/2529-maximum-count-of-positive-integer-and-negative-integer.java
@@ -1,6 +1,6 @@
 class Solution {
     public int maximumCount(int[] nums) {
-        int n = 0, p = 0;
+        short n = 0, p = 0;
         for (int i = 0; i < nums.length; i++) {
             if (nums[i] > 0) {
                 p++;
@@ -9,7 +9,7 @@ class Solution {
                 n++;
             }
         }
-        return Math.max(p, n);
+        return (int)Math.max(p, n);
 
     }
 }
